+++ podcasts/Strings+Generated.swift
@@ -944,6 +944,12 @@ internal enum L10n {
   internal static var eacInformationalBannerListeningHistoryIconAccessibility: String { return L10n.tr("Localizable", "eac_informational_banner_listening_history_icon_accessibility", fallback: "Listening History Icon") }
   /// Encourage Account Creation: title for the banner in the listening history
   internal static var eacInformationalBannerListeningHistoryTitle: String { return L10n.tr("Localizable", "eac_informational_banner_listening_history_title", fallback: "Keep track of what you’ve played") }
+  /// Encourage Account Creation: description for the banner in the playlists view
+  internal static var eacInformationalBannerPlaylistsDescription: String { return L10n.tr("Localizable", "eac_informational_banner_playlists_description", fallback: "Create a free account to sync your playlists on any device.") }
+  /// Encourage Account Creation: accessibility label for the banner in the playlists view
+  internal static var eacInformationalBannerPlaylistsIconAccessibility: String { return L10n.tr("Localizable", "eac_informational_banner_playlists_icon_accessibility", fallback: "Playlists Icon") }
+  /// Encourage Account Creation: title for the banner in the playlists view
+  internal static var eacInformationalBannerPlaylistsTitle: String { return L10n.tr("Localizable", "eac_informational_banner_playlists_title", fallback: "Keep your playlists in sync") }
   /// Encourage Account Creation: description for the banner in the profile view
   internal static var eacInformationalBannerProfileDescription: String { return L10n.tr("Localizable", "eac_informational_banner_profile_description", fallback: "Create a free account to sync your shows and listen anywhere.") }
   /// Encourage Account Creation: accessibility label for the banner in the profile view
@@ -1897,7 +1903,7 @@ internal enum L10n {
   /// Settings to control sending of pocket casts offers notifications
   internal static var notificationsPocketCastOffers: String { return L10n.tr("Localizable", "notifications_pocket_cast_offers", fallback: "Pocket Casts Offers") }
   /// Notification body for recommendations trending message
-  internal static var notificationsRecommendationsTrendingBody: String { return L10n.tr("Localizable", "notifications_recommendations_trending_body", fallback: "Check out what everyone else is listening this week.") }
+  internal static var notificationsRecommendationsTrendingBody: String { return L10n.tr("Localizable", "notifications_recommendations_trending_body", fallback: "Check out what everyone else is listening to this week.") }
   /// Notification title for recommendations trending message
   internal static var notificationsRecommendationsTrendingTitle: String { return L10n.tr("Localizable", "notifications_recommendations_trending_title", fallback: "Trending this week") }
   /// Notification body for recommendations you might like message